requirements.txt
sqlalchemy==1.1.5
requests==2.13.0
PyGithub==1.32
oauth2client==4.0.0
google-api-python-client==1.6.2
configparser==3.5.0